app/requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
exifread>=3.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
insightface>=0.7.3
onnxruntime>=1.16.0
reverse_geocoder>=1.5.1
